Docs: Add warnings section to README
This commit adds a "Warnings" section to the README.md file. This section informs users about the potential risks of flashing kernels, emphasizing the importance of backups and compatibility checks.

Additionally, this commit enables code minification and resource shrinking for release builds in the `app/build.gradle.kts` file by setting `isMinifyEnabled` and `isShrinkResources` to `true`.